+++ modules/python_engine/builtin/include/py_isequalBuiltin.hpp
@@ -1,13 +1,20 @@
 //=============================================================================
+// Copyright (c) 2016-present Allan CORNET (Nelson)
+//=============================================================================
+// This file is part of the Nelson.
+//=============================================================================
+// LICENCE_BLOCK_BEGIN
+// SPDX-License-Identifier: LGPL-3.0-or-later
+// LICENCE_BLOCK_END
+//=============================================================================
 #pragma once
 //=============================================================================
 #include "ArrayOf.hpp"
-#include "Evaluator.hpp"
 //=============================================================================
-namespace Nelson::ElementaryFunctionsGateway {
+namespace Nelson::Python_engineGateway {
 //=============================================================================
 ArrayOfVector
-isequalBuiltin(Evaluator* eval, int nLhs, const ArrayOfVector& argIn);
+py_isequalBuiltin(int nLhs, const ArrayOfVector& argIn);
 //=============================================================================
 } // namespace Nelson
 //=============================================================================